property cannot be deleted under pending

+++ src/main/java/com/theateam/waaprojectspringbackend/service/PropertyService.java
@@ -55,6 +55,11 @@ public class PropertyService {
     }
 
     public void delete(Long propertyId) {
+        Property property = propertyRepo.findById(propertyId).orElseThrow();
+        if (property.getStatus().equals(PropertyStatus.STATUS_PENDING)) {
+            throw new RuntimeException("Property cannot be deleted");
+        }
+
         propertyRepo.deleteById(propertyId);
     }
 }
